move part to lib maplibre
